Во временном классе бочки тип данных изменен на double. Добавлен помник. Добавлены юнит тесты.

+++ src/main/java/Barrel.java
@@ -1,9 +1,9 @@
 public class Barrel {
-    public int size;
+    public double size;
     public String material;
     public String filling;
 
-    public Barrel(int size, String material, String filling){
+    public Barrel(double size, String material, String filling){
         this.size = size;
         this.material = material;
         this.filling = filling;
